CollidingSpheres compiles. Updates in project
Improvements in python files
vectorIO doesn't have casts anymore
some ints changed to unsigneds for compatibility with vector::size()
_itoa substituted by to_string

+++ IOLib/source/vectorIO.cpp
@@ -15,7 +15,7 @@ istream & operator >> (istream & stream, Vector3D & v){
 
 
 ofstream & operator << (ofstream & stream, const Vector3D & v){
-	stream << static_cast<long double>(v.x()) << "    " << static_cast<long double>(v.y()) << "    " << static_cast<long double>(v.z()) << endl;
+	stream << v.x() << "    " << v.y() << "    " << v.z() << endl;
 	return stream;
 }
